% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DataR_datasetViewer.R
\name{create_DT_from_dataframe}
\alias{create_DT_from_dataframe}
\title{Create dataset viewer using DataTable}
\usage{
create_DT_from_dataframe(d, transpose, columns_to_display)
}
\arguments{
\item{d}{data.frame to be displayed}

\item{transpose}{boolean, TRUE to transpose the table view}

\item{columns_to_display}{character column names to display}
}
\value{
DataTable HTML
}
\description{
Create dataset viewer using DataTable
}
