% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{insertDashboardHeader}
\alias{insertDashboardHeader}
\title{Wrap ui elements in dropdown tag
so they can be placed in the shiny dashboard header}
\usage{
insertDashboardHeader(x, show)
}
\arguments{
\item{x}{}

\item{show}{boolean TRUE shows item, FALSE hides item}
}
\value{
HTML tag
}
\description{
Wrap ui elements in dropdown tag
so they can be placed in the shiny dashboard header
}
