% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{insurance}
\alias{insurance}
\title{Prices of 25,000 motor insurance policies}
\format{
A data frame with 25000 rows and 17 variables:
\describe{
  \item{price}{price paid for the policy}
  \item{age}{the customer's age}
  \item{experience}{the number of years the customer has held their driving licence}
  \item{licence}{the type of driving licence}
  \item{NCD}{the number of years' no claims discount}
  \item{PostcodeArea}{the UK postcode area of the customer's address}
  \item{postcode_cat}{numerical postcode risk category}
  \item{make}{vehicle make}
  \item{value}{the value of the vehicle}
  \item{vehicle_age}{the age of the vehicle}
  \item{ownership}{how many years the customer has owned the vehicle}
  \item{vehicle_cat}{numerical vehicle risk category}
  \item{vehicle_kept}{where the vehicle is kept overnight}
  \item{mileage}{the number of miles driven per year}
  \item{use}{what the vehicle is used for}
  \item{claims}{the number of prior claims the customer has made}
  \item{train_test}{training rows are 0, test rows are 1}
}
}
\usage{
insurance
}
\description{
A dataset containing the prices and other attributes of 25,000
 motor insurance customers. The variables are as follows:
}
\keyword{datasets}
