% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{lucidum}
\alias{lucidum}
\title{Launch lucidum}
\usage{
lucidum(
  data = NULL,
  BoostaR_models = NULL,
  GlimmaR_models = NULL,
  kpi_spec = NULL,
  filter_spec = NULL,
  feature_spec = NULL,
  specification_path = NULL,
  show_DataR = TRUE,
  show_ChartaR = TRUE,
  show_MappaR = TRUE,
  show_BoostaR = TRUE,
  show_GlimmaR = TRUE,
  show_DevelopaR = FALSE,
  starting_tab = "ChartaR",
  starting_response = NULL,
  sidebar_width = 280,
  num_threads = -1,
  title = NULL
)
}
\arguments{
\item{data}{data.frame or data.table to analyse in lucidum.}

\item{BoostaR_models}{character, path to .rds file containing BoostaR_models.}

\item{GlimmaR_models}{character, path to .rds file containing GlimmaR_models.}

\item{kpi_spec}{character, path to kpi specification file.}

\item{filter_spec}{character, path to filter specification file.}

\item{feature_spec}{character, path to feature specification file.}

\item{specification_path}{character, path to feature specification folder}

\item{show_DataR}{logical, TRUE (not default) will show the DataR menu item.}

\item{show_ChartaR}{logical, TRUE (default) will show the ChartaR menu item.}

\item{show_MappaR}{logical, TRUE (default) will show the MappaR menu item.}

\item{show_BoostaR}{logical, TRUE (default) will show the BoostaR menu item.}

\item{show_GlimmaR}{logical, TRUE (default) will show the GlimmaR menu item.}

\item{show_DevelopaR}{logical, TRUE (not default) will show the ShinyAce console.}

\item{starting_tab}{character, name of tab to show on startup.}

\item{starting_response}{character, name of response column to show on startup.}

\item{sidebar_width}{integer, width of the sidebar.}

\item{num_threads}{integer, number of threads for data.table and lightgbm (default -1 means max threads)}

\item{title}{character, text to show in the app header.}
}
\description{
Launch lucidum
}
\examples{
# Launch lucidum with the supplied demo dataset called "insurance"

lucidum(insurance, starting_response = 'price')

}
